+++ apps/web/src/features/dashboard/components/appHeaderLinks.ts
@@ -1,0 +1,11 @@
+export const PRACTICE_LINKS = [
+  { to: '/daily', label: 'デイリーチャレンジ' },
+  { to: '/practice/code-doctor', label: 'コードドクター' },
+  { to: '/practice/mini-projects', label: 'ミニプロジェクト' },
+  { to: '/practice/code-reading', label: 'コードリーディング' },
+] as const
+
+export const TOP_NAV_LINKS = [
+  { to: '/base-nook', label: 'ベースヌック', pathPrefix: '/base-nook' },
+  { to: '/profile', label: 'プロフィール', pathPrefix: '/profile' },
+] as const
